\name{H7N9}
\alias{H7N9}
\docType{data}
\title{H7N9 data set}
\description{
Text statuses about H7N9 bird flu in Weibo.
}
\usage{data(H7N9)}
\format{
  A data frame with 8 columns.
  \describe{
    \item{\code{MID}}{Weibo ID.}
    \item{\code{Author}}{Nick name of the author.}
    \item{\code{Weibo}}{Weibo content.}
    \item{\code{Forward}}{Forward content.}
    \item{\code{Time_Weibo}}{Time of this weibo.}
    \item{\code{Time_Search}}{Time of this search.}
    \item{\code{Count_Forward}}{Count of forwards.}
    \item{\code{Count_Reply}}{Count of replys.}
  }
}

\author{
  Jian Li <\email{rinds.book@gmail.com}>
}
\references{
  \url{http://weibo.com/}
}
\keyword{datasets}
